sovran_v2/requirements.txt
httpx>=0.27.0
python-dotenv>=1.0.0
pytest>=8.0.0
pytest-asyncio>=0.23.0
websocket-client>=1.7.0
